lenght is now float type

+++ db/migrate/20180925200613_create_routes.rb
@@ -1,7 +1,7 @@
 class CreateRoutes < ActiveRecord::Migration[5.2]
   def change
     create_table :routes do |t|
-      t.integer :length
+      t.float :length
       t.references :user, foreign_key: true
 
       t.timestamps
